Alot of updates
This commit consists of a working web app that has the following:
1. Designed with base style and phase 1 layout.
2. Create page that consits of 3 fields to create a "student".
3. On home page load, there will be a table that gets generated from
the controller by calling a function inside the model (storage.php)
compiling the results.
4.  The table consits of all the information regarding the student
table in the database, as well as a link titled "Edit" after each row.
Clicking this will bring you to the edit page where you will be able to
edit that particular students' profile.